octofit-tracker/backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
django-allauth==0.51.0
django-cors-headers==3.14.0
dj-rest-auth==2.2.5
djongo==1.3.6
pymongo==3.12.0
sqlparse==0.2.4
